Move CreateTextLoader helper to TestMocks class
